fix(transact): keep upsert absence undecidable when the store failed to load
`generate_upsert_deletions` treated a missing range provider as proof that no
base index exists, so every subject was decided by novelty alone. That holds at
genesis, but the ledger manager treats a binary-store load failure as non-fatal
("Failed to load binary store, continuing without"), which leaves an indexed
ledger (`t > 0`) with no provider attached. There the pre-check reported every
subject absent, skipped it, and committed zero retractions -- old values
silently survived alongside the new ones. Before the skip existed that state
failed loudly at `range_with_overlay`.

- stage.rs: require `t == 0` alongside the missing provider, mirroring
  `fluree_db_core::range`, which treats a missing provider as an empty index
  only at genesis and errors otherwise. A `t > 0` ledger with no provider now
  falls through to the per-predicate query, which surfaces the load failure.
- stage.rs: drop the `store.sid_to_iri` fallback. The store's namespace table
  is a subset of the snapshot's -- the index root is a materialized cache at
  `index_t` and `sync_store_and_snapshot_ns` reconciles its codes back on load
  -- so it can never decode a code the snapshot could not. Collapse the
  closure's no-store arm to `false`, which is all `!base_index_absent` could
  ever evaluate to under the caller's gate.

Tests: the per-subject-namespace case now covers a subject committed after the
last index build, where the snapshot decodes the namespace code but the store
conclusively misses, so only the novelty backstop prevents a wrong skip. That
is the one direction the fail-open removal could lose a retraction in, and it
is visible in the results; verified by neutralizing the backstop, which fails
exactly that assertion. The previously covered subject was already indexed and
never reached the branch.

Move the span-capture assertion out of the grouped binary into its own
`it_upsert_skip_fires` test, per the standalone rule in
docs/contributing/tests.md: tracing's callsite-interest cache is process-global
and the capture thread-local, so a sibling test in the same process can leave
the event undelivered.

Docs: namespace-allocation no longer suggests "setting HostPlusN(n) on the
ledger" -- there is no user-facing knob, the bulk-import preflight is the only
thing that selects it, and the mode is immutable after genesis.
