+++ docs/contributing/tests.md
@@ -92,7 +92,8 @@ binary. No new `[[test]]` entry is needed unless you are creating a new group.
   so this only matters for bare `cargo test`.)
 - Tests whose **assertions are about instrumentation** — presence/absence of
   tracing events captured via a thread-local `span_capture` subscriber (e.g.
-  `it_cyclic_bgp_probe`, `it_minmax_fast_path_fired`). Tracing's
+  `it_cyclic_bgp_probe`, `it_minmax_fast_path_fired`,
+  `it_upsert_skip_fires`). Tracing's
   callsite-interest cache is process-global and the capture is thread-local,
   so these assertions are only reliable when the test is alone in its
   process. (Same caveat as above: bites bare `cargo test` only.)
